+++ CHANGELOG.md
@@ -53,6 +53,7 @@
 - Add function in `PatientRepository` to check if patient is eligible for reassignment
 - Bump eclipse jgit to v6.9.0.202403050737-r
 - Update GH Actions `checkout` to v4
+- Add DB query to update patient reassignment status
 
 ### Fixes